chatbot/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
chromadb==0.4.18
python-dotenv==1.0.0
pypdf==3.17.4
tiktoken==0.5.2
markdown==3.5.1
python-markdown==3.5.1
